booking_service/pyproject.toml
[project]
name = "booking-service"
version = "0.1.0"
description = "Booking service for Hotel System"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[all]",
    "sqlalchemy",
    "asyncpg",
    "psycopg2-binary",
    "pydantic-settings",
    "python-dotenv"
]

[tool.setuptools]
package-dir = {"" = "app"}
packages = ["api", "config", "db", "models", "service"] 